Update codeclimate formatter to never show source
The codeclimate formatter breaks if the show source option is activated. This option would previously insert the source line, where the error occured as raw text into the json. This leads to parsing errors, when reading the generated json output. With the proposed changes, the source line will never be inserted.